feat(tabs-on-kit-page): create asset tab on kit page
