+++ app/components/assets/assets-index/filters.tsx
@@ -23,7 +23,7 @@ import { resolveTeamMemberName } from "~/utils/user";
 import { AdvancedFilteringAndSorting } from "./advanced-asset-index-filters-and-sorting";
 import { ConfigureColumnsDropdown } from "./configure-columns-dropdown";
 
-export const ASSET_INDEX_SORTING_OPTIONS = {
+export const ASSET_SORTING_OPTIONS = {
   title: "Name",
   createdAt: "Date created",
   updatedAt: "Date updated",
@@ -52,7 +52,7 @@ export function AssetIndexFilters({
           "left-of-search": <StatusFilter statusItems={AssetStatus} />,
           "right-of-search": (
             <SortBy
-              sortingOptions={ASSET_INDEX_SORTING_OPTIONS}
+              sortingOptions={ASSET_SORTING_OPTIONS}
               defaultSortingBy="createdAt"
             />
           ),
