+++ app/components/kits/booking-actions-dropdown.tsx
@@ -37,7 +37,7 @@ export default function BookingActionsDropdown() {
       disabled,
       label: "Create new booking",
       icon: "bookings",
-      to: "create-new-booking",
+      to: "assets/create-new-booking",
     },
     {
       indexType: "kit",
@@ -45,7 +45,7 @@ export default function BookingActionsDropdown() {
       label: "Add to existing booking",
       icon: "booking-exist",
       disabled,
-      to: `/kits/${kit.id}/add-to-existing-booking`,
+      to: `/kits/${kit.id}/assets/add-to-existing-booking`,
     },
   ] as BookLink[];
