+++ app/routes/_layout+/kits.$kitId.assets.create-new-booking.tsx
