+++ app/routes/_layout+/kits.$kitId.assets.release-custody.tsx
@@ -45,7 +45,7 @@ export async function loader({ context, request, params }: LoaderFunctionArgs) {
       request,
     });
     if (!kit.custody) {
-      return redirect(`/kits/${kitId}`);
+      return redirect(`/kits/${kitId}/assets`);
     }
 
     /**
@@ -128,7 +128,7 @@ export async function action({ context, request, params }: ActionFunctionArgs) {
       senderId: userId,
     });
 
-    return redirect(`/kits/${kitId}`);
+    return redirect(`/kits/${kitId}/assets`);
   } catch (cause) {
     const reason = makeShelfError(cause, { userId, kitId });
     return json(error(reason), { status: reason.status });
