+++ app/routes/_layout+/locations.$locationId.add-assets.tsx
@@ -13,7 +13,7 @@ import {
 } from "~/atoms/list";
 import { AssetImage } from "~/components/assets/asset-image/component";
 import { AssetStatusBadge } from "~/components/assets/asset-status-badge";
-import { ASSET_INDEX_SORTING_OPTIONS } from "~/components/assets/assets-index/filters";
+import { ASSET_SORTING_OPTIONS } from "~/components/assets/assets-index/filters";
 import { StatusFilter } from "~/components/booking/status-filter";
 import { Form } from "~/components/custom-form";
 import DynamicDropdown from "~/components/dynamic-dropdown/dynamic-dropdown";
@@ -378,7 +378,7 @@ export default function AddAssetsToLocation() {
             "left-of-search": <StatusFilter statusItems={AssetStatus} />,
             "right-of-search": (
               <SortBy
-                sortingOptions={ASSET_INDEX_SORTING_OPTIONS}
+                sortingOptions={ASSET_SORTING_OPTIONS}
                 defaultSortingBy="createdAt"
               />
             ),
